pyproject.toml
[tool.poetry]
name = "aitm-data-collection"
version = "0.1.0"
description = ""
authors = ["Sondre Halvorsen <Sondre.Halvorsen@nav.no>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
